requirements.txt
Django==2.0.1
django-cleanup==2.1.0
django-dynamic-preferences==1.5
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
Markdown==2.6.11
persisting-theory==0.2.1
Pillow==5.0.0
pypugjs==5.1.3
pytz==2017.3
six==1.11.0
